Testing ScrollContainer
Tried testing the constraints of the scroll container when the vertical scroll is not disabled. When it is disabled, with the way the Hbox and Vboxes are, adding too many textureRects to the vboxes makes the scroll container grow past the viewport, which is no. I want the textureRects to shrink as you add more in a vbox. This branch is ui-skill-tree.