Made reloading
Can reload using the R key (physical) or by using left click when there are no more bullets in the chamber. Set up a string variable gun_name in the player script that changes when the gun changes. Assuming that a gun class or duck typing will be used with the same variable names. Done so the number of bullets in the chamber and reserve are kept between levels. Is this messy? Probably, yeah. Too bad!